Update lerpmonitor.sp
Hello here is an update to use the new version of sourcemod. But there is an error on line 161 and 220. (error 035: argument type mismatch (argument 2)). To see the details of the changes see the comments added to the code. If I can update the code completely, I will send a new message.